Combine profit.js and easytxf.js
There is a major change about how opening date is determined. Previously, if a large sell order consumes several buy orders, the opening date of all the sales that this sell order produces are the same and it's the date of the first consumed buy order. This is fixed in this commit